% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_water_yr_mth.R
\name{append_water_yr_mth}
\alias{append_water_yr_mth}
\title{Append water year and water month to a data frame}
\usage{
append_water_yr_mth(dat)
}
\arguments{
\item{dat}{A data frame containing numeric 'month' and 'year' columns.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
This function appends five (5) columns to the input data set based on the
month and year columns: \code{water_month}, \code{month_name}, \code{season},
\code{water_yr}, and \code{water_mth.}
}
\details{
The following columns are appended to the input data frame:
\itemize{
\item \code{water_month}: Numeric. Water month Oct-Sep (1-12).
\item \code{month_name}: Factor. Abbreviated month name with levels corresponding to the calendar year (Jan-Dec).
\item \code{season}: Character. season based on water year (DJF, MAM, JJA, SON)
\item \code{water_yr}: Numeric. Year corresponding to water year
\item \code{water_mth}: Factor. Abbreviated month name with levels corresponding to the water year (Oct-Sep).
}
}
\examples{
library(climateAnalyzeR)

# Import monthly departures and append water year and water month
import_departure("canyonlands_theneck", 1965, 2025) |>
  append_water_yr_mth()

}
