% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_prcp}
\alias{convert_prcp}
\title{convert_prcp: Convert precipitation data from imperial to metric}
\usage{
convert_prcp(dat)
}
\arguments{
\item{dat}{data frame or tibble}
}
\value{
dataframe
}
\description{
convert_prcp: Convert precipitation data from imperial to metric
}
