% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_temp}
\alias{convert_temp}
\title{convert_temp: Convert temperature data from imperial to metric}
\usage{
convert_temp(dat)
}
\arguments{
\item{dat}{data frame or tibble}
}
\value{
dataframe
}
\description{
convert_temp: Convert temperature data from imperial to metric
}
