% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{glue_mths}
\alias{glue_mths}
\title{glue_mths: Glue Months}
\usage{
glue_mths(mths)
}
\arguments{
\item{mths}{Vector of month names or abbreviations}
}
\value{
string
}
\description{
Returns a string of months with comma's between each month.
}
