% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_annual.R
\name{import_annual}
\alias{import_annual}
\title{Import annual temperature and precipitation data}
\usage{
import_annual(
  station_id,
  start_year,
  end_year = NULL,
  screen_blanks = "true",
  remove_missing = TRUE,
  convert = FALSE
)
}
\arguments{
\item{station_id}{The character string of the \emph{station_id} field from
\code{\link{stations}}.}

\item{start_year}{The four digit number of the first year of interest.}

\item{end_year}{The four digit number of the last year of interest. Default
is NULL. If NULL, current year will be used.}

\item{screen_blanks}{A character string stating if years with 15 or more
missing values will be screened out and the data replaced with NA.
Options are 'true' or 'false', in lower case not upper case. This object
is included directly into the URL to retrieve the data from
ClimateAnalyzer.org.}

\item{remove_missing}{Logical. If TRUE, columns that tally missing values are
excluded from result. If FALSE, these columns are included in the result.
Default is TRUE.}

\item{convert}{Logical. If TRUE, data are precipitation and temperature
values are converted to metric. These converted values are included as
additional columns in the data frame denoted by "_mm" or "_C". Default is
FALSE.}
}
\value{
A \code{\link[tibble]{tibble}}.
}
\description{
This function imports annual temperature and precipitation data from
\href{http://www.climateanalyzer.org/}{ClimateAnalyzer.org} into R.
}
\examples{
library(climateAnalyzeR)

# Import annual temperature and precipitation data
import_annual(station_id = "zion_np", start_year = 1980)

# Import annual temperature and precipitation data and convert values to metric
import_annual(station_id = "zion_np", start_year = 1980, end_year = 2020,
              convert = TRUE)

}
\seealso{
The \code{\link{import_data}} wrapper function.
}
