% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_below_above_temp.R
\name{import_below_above_temp}
\alias{import_below_above_temp}
\title{Import count of days with temperatures below and above user set temps}
\usage{
import_below_above_temp(
  station_id,
  start_year,
  end_year,
  tmin_temp,
  tmax_temp,
  station_type,
  year = NULL
)
}
\arguments{
\item{station_id}{The character string of the \emph{station_id} field from
\code{\link{stations}}.}

\item{start_year}{The four digit number of the last year of interest.}

\item{end_year}{The four digit number of the last year of interest.}

\item{tmin_temp}{One to three digit number, including - for temperatures
below zero, for the minimum temperature threshold in Fahrenheit}

\item{tmax_temp}{One to three digit number for the maximum temperature
threshold in Fahrenheit}

\item{station_type}{A character string for the type of weather station the
data are being pulled from. Options include "GHCN", "SNOTEL", and "RAWS".}

\item{year}{A character string for the year period. Options include "year"
for calendar year and "water_year". Default is "year". If using
"water_year"ensure that the start_year is the year prior to the period of
interest.}
}
\value{
A \code{\link[tibble:tibble]{name}}.
}
\description{
This function imports number of days annually with minimum and maximum daily
temperatures below and above user set temperatures.
column names.
\href{http://www.climateanalyzer.org/}{ClimateAnalyzer.org} into R.
}
\examples{
library(climateAnalyzeR)

import_below_above_temp(station_id = "tumacacori", start_year = 1991, end_year =
2020, tmin_temp = 20, tmax_temp = 100, station_type = "GHCN", year = "year")

}
\seealso{
The \code{\link{import_data}} wrapper function.
}
