% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_daily.R
\name{import_daily}
\alias{import_daily}
\title{Read daily weather data from ClimateAnalyzer.org into R}
\usage{
import_daily(station, start_year, end_year)
}
\arguments{
\item{station}{The character string of the station name.}

\item{start_year}{The four digit number of the first year of interest.}

\item{end_year}{The four digit number of the last year of interest.}
}
\value{
A \code{\link[tibble:tibble]{tibble}}.
}
\description{
Read daily weather data from ClimateAnalyzer.org into R
}
\examples{
import_daily("arches", 2010, 2020)

}
