% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_monthly.R
\name{import_monthly}
\alias{import_monthly}
\title{Import monthly temperature and precipitation data}
\usage{
import_monthly(
  station_id,
  start_year,
  end_year,
  month = "all",
  convert = FALSE
)
}
\arguments{
\item{station_id}{The character string of the station ID.}

\item{start_year}{The four digit number of the first year of interest.}

\item{end_year}{The four digit number of the last year of interest.}

\item{month}{The month number (i.e., 1 for January through 12 for December)
or 'all' for all months.}

\item{convert}{Logical. If TRUE, data are precipitation and temperature
values are converted to metric. These converted values are included as
additional columns in the data frame denoted by "_mm" or "_C". Default is
FALSE.}
}
\value{
A \code{\link[tibble:tibble]{tibble}}.
}
\description{
This function imports monthly temperature and precipitation data from
\href{http://www.climateanalyzer.org/}{ClimateAnalyzer.org} into R.
}
\examples{
# Import monthly precipitation and temperature data
import_monthly('canyonlands_theneedle', 2000, 2010)

# Import monthly precipitation and temperature data for the month of June and
# convert values to metric
import_monthly('canyonlands_theneedle', 2000, 2010, month = 6, convert = TRUE)

}
\seealso{
The \code{\link{import_data}} wrapper function.
}
