% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_water_balance.R
\name{import_water_balance}
\alias{import_water_balance}
\title{Import water balance data from ClimateAnalyzer.org into R}
\usage{
import_water_balance(
  station,
  start_year,
  end_year,
  table_type,
  pet_type,
  soil_water,
  forgiving
)
}
\arguments{
\item{station}{The character string of the station name.}

\item{start_year}{The four digit number of the first year of interest.}

\item{end_year}{The four digit number of the last year of interest.}

\item{table_type}{A character string for 'daily' or 'monthly' data.}

\item{pet_type}{A character string for the model type. Options are 'Hamon' or
'Penman_Montieth'.}

\item{soil_water}{a number (integer) for the soil depth in milimeters. If you are not
sure use 100.}

\item{forgiving}{A string for the tolorance of the model. Options are 'no',
'mild' and 'very'.}
}
\value{
A \code{\link[tibble]{tibble}}.
}
\description{
Import water balance data from ClimateAnalyzer.org into R
}
\examples{
import_water_balance("arches", 2015, 2020, table_type = "monthly",
                     soil_water = 100, pet_type = "hamon",
                     forgiving = "very")
# Adjust the soil water capacity to 50 mm and use the Penman_Montieth model.
import_water_balance("arches", 2015, 2020, table_type = "monthly",
                     soil_water = 50, pet_type = "Penman_Montieth",
                     forgiving = "very")
}
