% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{missing_arg}
\alias{missing_arg}
\title{missing_arg: Simple function to create an error message when arguments are missing}
\usage{
missing_arg(my_arg, my_mess)
}
\arguments{
\item{my_arg}{Function arguments}

\item{my_mess}{Error message}
}
\description{
missing_arg: Simple function to create an error message when arguments are missing
}
