% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{number_contraction}
\alias{number_contraction}
\title{number_contraction: Number Contraction.}
\usage{
number_contraction(x)
}
\arguments{
\item{x}{vector of numbers}
}
\value{
string
}
\description{
Prints the appropriate number contraction for a given number.
}
