% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pull_csv}
\alias{pull_csv}
\title{pull_csv: Pull comma delimited data}
\usage{
pull_csv(my_url, skip)
}
\arguments{
\item{my_url}{URL}

\item{skip}{Lines to skip}
}
\value{
dataframe
}
\description{
pull_csv: Pull comma delimited data
}
