% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pull_monthly}
\alias{pull_monthly}
\title{Function to import monthly data}
\usage{
pull_monthly(station_id, start_year, end_year, month, table_type, norm_per)
}
\description{
Function to import monthly data
}
