% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pull_xml}
\alias{pull_xml}
\title{pull_xml: Child function to scrape HTML tables}
\usage{
pull_xml(my_url, skip)
}
\arguments{
\item{my_url}{URL}

\item{skip}{rows to skip}
}
\value{
dataframe
}
\description{
pull_xml: Child function to scrape HTML tables
}
