% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderSummaries.R
\name{renderSummary}
\alias{renderSummary}
\title{Render reports}
\usage{
renderSummary(
  station_id,
  station_name = NULL,
  my_year,
  report = "water_year",
  my_dir = NULL
)
}
\arguments{
\item{station_id}{The station identification name for the weather station of
interest.}

\item{station_name}{Optional. The name of the station to display in the report.
Default is NULL. If left as NULL the station name will be pulled using the
stations() function.}

\item{report}{Type of report to render. Currently the only option is a water
year ("water_year") report.}

\item{my_dir}{String of direectory path to save repor to. If NULL, the report
is saved to the your desktop.}

\item{year}{The four-digit for the year of interest.}
}
\value{
A PDF.
}
\description{
This funciton will render a PDF report from an RMarkdown (Rmd) script.
}
\examples{
# Render a report for Arches National Park for the 2020 water year.
renderSummary("Arches National Park", 2020, "ARCH", report = "Water Year")
}
