% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_figure.R
\name{annual_figure}
\alias{annual_figure}
\title{Line Figure with Area}
\usage{
annual_figure(
  x_var,
  y_var,
  normal,
  reference_period,
  area_color,
  line_color,
  my_title,
  my_ylab
)
}
\arguments{
\item{x_var}{A character string of hte x-variable.}

\item{y_var}{A character string of the y-variable.}

\item{normal}{A number for the annual normal temperature or precipitation. Typically derived from \code{\link[climateAnalyzeR::normals]{normals}}.}

\item{reference_period}{Character string for the reference period used for normals.}

\item{area_color}{A character string for the color of the area. See \code{\link[grDevices::colors]{colors}} for a list of color names.}

\item{line_color}{A character string for the line color.}

\item{my_title}{A character string for the title.}

\item{my_ylab}{A character string from the y-axis.}
}
\value{
A ggplot object.
}
\description{
This function produces a standardized line graph with colored area under the line. This figure is used in RMarkdown scrips.
}
\examples{
# TODO: use prcp data
}
