% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_figure.R
\name{annual_figure}
\alias{annual_figure}
\title{Annual Figure}
\usage{
annual_figure(
  x_var,
  y_var,
  my_year,
  normal,
  reference_period,
  area_color,
  line_color,
  my_title,
  my_ylab
)
}
\arguments{
\item{x_var}{A character string of the x-variable. This is intended to be a
vector of years.}

\item{y_var}{A character string of the y-variable. This is intended to be a
vector of measurements (e.g., PRCP or TEMP).}

\item{my_year}{A number for the your of interest.}

\item{normal}{A number for the annual normal temperature or precipitation.
Typically from \code{\link[climateAnalyzeR::normals]{normals}}.}

\item{reference_period}{Character string for the reference period used for
normals (e.g., "1981-2010").}

\item{area_color}{A character string for the color of the area. See
\code{\link[grDevices::colors]{colors}} for a list of color names.}

\item{line_color}{A character string for the line color.}

\item{my_title}{A character string for the title.}

\item{my_ylab}{A character string from the y-axis.}
}
\value{
A ggplot object.
}
\description{
This function produces a standardized line graph with colored area under the
line and a trend line. This figure is used for annual measurements (e.g.,
precipitation or temperature) in RMarkdown scrips to produce reports.
}
\examples{
# Import annual data from ClimateAnalyzer.org
dat <-  import_data("annual_wx", "canyonlands_theneck", 1980, 2020,
                    remove_missing = FALSE)
# Import NOAA calculated normals
dat.30yr.prcp <- normals(my_stations = "canyonlands_theneck") \%>\%
  dplyr::filter(element == "PRCP" & month == "Annual")
# Plot data
annual_figure(x_var = dat$year, y_var = dat$prcp_wy, my_year = 2018,
              normal = dat.30yr.prcp$value, reference_period = "1981-2010",
              area_color = "green3", line_color = "darkgreen",
              my_title = "Annual Precipitation, Island in the Sky, Canyonlands NP",
              my_ylab = "PRCP (inches)")

}
