% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{annual_figure}
\alias{annual_figure}
\title{Annual Figure}
\usage{
annual_figure(
  x_var,
  y_var,
  normal,
  reference_period,
  area_color,
  line_color,
  my_title,
  my_ylab
)
}
\arguments{
\item{x_var}{A character string of the x-variable. This is intended to be a
vector of years.}

\item{y_var}{A character string of the y-variable. This is intended to be a
vector of measurements (e.g., PRCP or TEMP).}

\item{normal}{A number for the annual normal temperature or precipitation.
Typically from \code{\link{normals}}.}

\item{reference_period}{Character string for the reference period used for
normals (e.g., "1981-2010").}

\item{area_color}{A character string for the color of the area. See
\code{\link[grDevices]{colors}} for a list of color names.}

\item{line_color}{A character string for the line color.}

\item{my_title}{A character string for the title.}

\item{my_ylab}{A character string from the y-axis.}
}
\value{
A ggplot object.
}
\description{
This function produces a standardized line graph with colored area under the
line and a trend line. This figure is used for annual measurements (e.g.,
precipitation or temperature) in RMarkdown scrips to produce reports.
}
\examples{
library(climateAnalyzeR)

# Set my station
my_station <- "colorado_nm"

# Import annual data from ClimateAnalyzer.org
dat <-  import_annual(station_id =  my_station, start_year = 1980,
                      end_year = 2020)

# Import NOAA calculated normals
prcp_30yr <- normals(station_id = my_station) |>
  dplyr::filter(element == "PRCP" & month == "Annual")

# Plot data
annual_figure(x_var = dat$year, y_var = dat$prcp_wy, my_year = 2019,
              normal = prcp_30yr$value, reference_period = "1981-2010",
              area_color = "green3", line_color = "darkgreen",
              my_title = "Annual Precipitation, Colorado National Monument",
              my_ylab = "PRCP (inches)")

}
