% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_annual.R
\name{import_annual}
\alias{import_annual}
\title{Import annual weather data from ClimateAnalyzer.org into R}
\usage{
import_annual(station, start_year, end_year, screen_blanks = "true")
}
\arguments{
\item{station}{The character string of the station name.}

\item{start_year}{The four digit number of the first year of interest.}

\item{end_year}{The four digit number of the last year of interest.}

\item{screen_blanks}{A character string stating if years with 15 or more
missing values will be screened out and the data replaced with NA.
Options are 'true' or 'false'.}
}
\value{
A \code{\link[tibble]{tibble}}.
}
\description{
Import annual weather data from ClimateAnalyzer.org into R
}
\examples{
import_annual("arches", 1980, 2020) \%>\%
  dplyr::select(-dplyr::contains("missing"))
}
\seealso{
\code{\link{import_daily}}, \code{\link{import_monthly}}
}
