% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_daily.R
\name{import_daily}
\alias{import_daily}
\title{Import daily weather data}
\usage{
import_daily(station_id, start_year, end_year = NULL, convert = FALSE)
}
\arguments{
\item{station_id}{The character string of the \emph{station_id} field from
\code{\link{stations}}.}

\item{start_year}{The four digit number of the last year of interest. Default
is NULL. If NULL, current year will be used.}

\item{end_year}{The four digit number of the last year of interest.}

\item{convert}{Logical. If TRUE, data are precipitation and temperature
values are converted to metric. These converted values are included as
additional columns in the data frame denoted by "_mm" or "_C". Default is
FALSE.}
}
\value{
A \code{\link[tibble:tibble]{name}}.
}
\description{
This function imports daily temperature and precipitation data from
\href{http://www.climateanalyzer.org/}{ClimateAnalyzer.org} into R.
}
\examples{
library(climateAnalyzeR)

# Import daily temperature and precipitation data
import_daily(station_id = "arches", start_year = 2010)

# Import daily temperature and precipitation data and convert values to metric
import_daily(station_id = "arches", start_year = 2010, end_year = 2020,
             convert = TRUE)

}
\seealso{
The \code{\link{import_data}} wrapper function.
}
