% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_daily.R
\name{import_daily}
\alias{import_daily}
\title{Import daily weather data}
\usage{
import_daily(station_id, start_year, end_year, convert = FALSE)
}
\arguments{
\item{station_id}{The character string of the station ID.}

\item{start_year}{The four digit number of the first year of interest.}

\item{end_year}{The four digit number of the last year of interest.}

\item{convert}{Logical. If TRUE, data are precipitation and temperature
values are converted to metric. These converted values are included as
additional columns in the data frame denoted by "_mm" or "_C". Default is
FALSE.}
}
\value{
A \code{\link[tibble:tibble]{tibble}}.
}
\description{
This function imports daily temperature and precipitation data from
\href{http://www.climateanalyzer.org/}{ClimateAnalyzer.org} into R.
}
\examples{
library(climateAnalyzeR)

# Import daily temperature and precipitation data
import_daily("hans_flat_rs", 2010, 2020)

# Import daily temperature and precipitation data and convert values to metric
import_daily("hans_flat_rs", 2010, 2020, convert = TRUE)

}
\seealso{
The \code{\link{import_data}} wrapper function.
}
