% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Read data from ClimateAnalyzer.org into R.}
\usage{
import_data(data_type, station_id, start_year, end_year = NULL, ...)
}
\arguments{
\item{data_type}{A character string for the type of data to import.
\itemize{
\item \emph{annual_wx}: Annual temperature and precipitation data.
See \code{\link{import_annual}} for details.
\item \emph{daily_wx}: Daily temperature and precipitation data.
See \code{\link{import_daily}} for details.
\item \emph{departure}: Monthly departure data.
See \code{\link{import_departure}} for details.
\item \emph{monthly_wx}: Monthly temperature and precipitation data.
See \code{\link{import_monthly}} for details.
\item \emph{water_balance}: Water balance data.
See \code{\link{import_water_balance}} for details.
}}

\item{station_id}{The character string of the \emph{station_id} field from
\code{\link{stations}}.}

\item{start_year}{The four digit number of the first year of interest.}

\item{end_year}{The four digit number of the last year of interest. Default
is NULL. If NULL, current year will be used.}

\item{...}{Other arguments to pass to the import functions.}
}
\value{
A \code{\link[tibble:tibble]{name}}.
}
\description{
This is a wrapper for data import functions in in the climateAnalyzeR package.
}
\examples{
library(climateAnalyzeR)

#-- Wrapper Function
# Import annual temperature and precipitation data
import_data("annual_wx", station_id = "zion_np", start_year = 1980)

# Import daily temperature and precipitation data
import_data("daily_wx", station_id = "arches", start_year = 2010,
            station_type = "GHCN")

# Import monthly departures
import_data("departure", station_id = "natural_bridges_nm", start_year = 2000)

# Import monthly precipitation and temperature data
import_data("monthly_wx", station_id = "canyonlands_theneedle",
            start_year = 2000)

# Import monthly water balance data using the Hamon model with soil water
# capacity set to 100.
import_data("water_balance", station_id = "bryce_canyon_np", start_year = 2015,
            end_year = 2020, table_type = "monthly", soil_water = 100,
            pet_type = "hamon", forgiving = "very")


}
\seealso{
See \code{\link{import_annual}}, \code{\link{import_daily}},
\code{\link{import_departure}}, \code{\link{import_monthly}}, and
\code{\link{import_water_balance}} for details.
}
