% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_extreme_temp.R
\name{import_extreme_temp}
\alias{import_extreme_temp}
\title{Import count of days with extreme temperatures}
\usage{
import_extreme_temp(
  station_id,
  start_year,
  end_year,
  station_type,
  year = NULL
)
}
\arguments{
\item{station_id}{The character string of the \emph{station_id} field from
\code{\link{stations}}.}

\item{start_year}{The four digit number of the last year of interest.}

\item{end_year}{The four digit number of the last year of interest.}

\item{station_type}{A character string for the type of weather station the
data are being pulled from. Options include "GHCN", "SNOTEL", and "RAWS".}

\item{year}{A character string for the year period. Options include "year"
for calendar year and "water_year". Default is "year". If using
"water_year"ensure that the start_year is the year prior to the period of
interest.}
}
\value{
A \code{\link[tibble:tibble]{name}}.
}
\description{
This function imports number of days annually with minimum temperatures below
the 5th percentile temperature and number of days above the 95th percentile,
mean consecutive days and max consecutive days of each. The 5th and 95th
percentile temps are relative to the time period chosen and are found in the
column names.
\href{http://www.climateanalyzer.org/}{ClimateAnalyzer.org} into R.
}
\examples{
library(climateAnalyzeR)

import_extreme_temp (station_id = "tumacacori", start_year = 1991, end_year =
2020, station_type = "GHCN", year = "year")

}
\seealso{
The \code{\link{import_data}} wrapper function.
}
