% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_monthly.R
\name{import_monthly}
\alias{import_monthly}
\alias{import_montly_wx}
\alias{import_monthly_depart}
\title{Import monthly data weather from ClimateAnalyzer.org into R}
\usage{
import_monthly(station, start_year, end_year, month, table_type, norm_per)

import_montly_wx(station, start_year, end_year, month = "all")

import_monthly_depart(
  station,
  start_year,
  end_year,
  month = "all",
  norm_per = "1981-2010"
)
}
\arguments{
\item{station}{The character string of the station name.}

\item{start_year}{The four digit number of the first year of interest.}

\item{end_year}{The four digit number of the last year of interest.}

\item{month}{The month number (i.e., 1 for January through 12 for December)
or 'all' for all months.}

\item{table_type}{A character sting for the type of table to produce.}

\item{norm_per}{A character string for the 30-year normalization period.}
}
\value{
A \code{\link[tibble]{tibble}}.
}
\description{
Import monthly data weather from ClimateAnalyzer.org into R
}
\examples{
import_monthly("arches", 2000, 2020, month = "all", table_type = "straight",
               norm_per = 'null')
import_montly_wx('arches', 2000, 2010)
import_monthly_depart('arches', 2000, 2010)
}
