% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_water_balance.R
\name{import_water_balance}
\alias{import_water_balance}
\title{Import water balance data}
\usage{
import_water_balance(
  station_id,
  start_year,
  end_year,
  table_type,
  pet_type,
  soil_water = 100,
  forgiving = "very"
)
}
\arguments{
\item{station_id}{The character string of the station ID.}

\item{start_year}{The four digit number of the first year of interest.}

\item{end_year}{The four digit number of the last year of interest.}

\item{table_type}{A character string for 'daily' or 'monthly' data.}

\item{pet_type}{A character string for the model type. Options are 'hamon' or
'Penman_Montieth'.}

\item{soil_water}{a number (integer) for the soil depth in millimeter.
Default is 100cm. If you are not sure use the default.}

\item{forgiving}{A string for the tolerance of the model. Options are 'no',
'mild' and 'very'. Default is 'very'.}
}
\value{
A \code{\link[tibble:tibble]{tibble}}.
}
\description{
This function imports daily or monthly water balance data from
\href{http://www.climateanalyzer.org/}{ClimateAnalyzer.org} into R.
}
\examples{
library(climateAnalyzeR)

# Import monthly water balance data using the Hamon model
import_water_balance("arches", 2015, 2020, table_type = "monthly",
                     soil_water = 100, pet_type = "hamon",
                     forgiving = "very")

# Import daily water balance data using the Penman Montieth model
import_water_balance("arches", 2015, 2020, table_type = "daily",
                     soil_water = 100, pet_type = "Penman_Montieth",
                     forgiving = "very")

}
\seealso{
The \code{\link{import_data}} wrapper function.
}
