% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normals.R
\name{normals}
\alias{normals}
\title{Import NOAA calculated normals from ClimateAnalyzer.org}
\usage{
normals(ref_period = "1981-2010", my_stations = NULL, tidy = TRUE)
}
\arguments{
\item{ref_period}{The string for 30-year reference period. The user can choose
"1971-2000" or "1981-2010" Default is "1981-2010".}

\item{my_stations}{Optional. A string to filter results by a station name.
Default is NULL. If NULL, a data frame of 30-year normals for all stations
on climateAnalyzer.org will be returned.}

\item{tidy}{Logical. Returns tidy data. Default is TRUE.
\describe{
\item{TRUE}{Returns a tidy data frame of temperature (tmax and tmin)
and precipitation data. This option dropps seasonl earlies/last
dates for temperatures and number of days above of below
temperatures.}
\item{FALSE}{Returns the raw dat from ClimateAnalyzer.org.}
}}
}
\value{
A \code{\link[tibble:tibble]{tibble}}.
}
\description{
This function imports 30-year normals from
\href{http://www.climateanalyzer.org/}{ClimateAnalyzer.org} into R.
}
\examples{
# 1981-2010 normals for Arches National Park
normals(my_stations = "arches")
# 1971-2000 normals for Arches National Park
normals(ref_period = "1971-2000", my_stations = "arches")
}
