% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normals.R
\name{normals}
\alias{normals}
\title{Import NOAA calculated normals}
\usage{
normals(ref_period = "1981-2010", station_id = NULL, tidy = TRUE)
}
\arguments{
\item{ref_period}{The string for 30-year reference period. The user can choose
"1971-2000", "1981-2010", or "1991-2020". Default is "1991-2020".}

\item{station_id}{Optional. The character string of the \emph{station_id}
field from \code{\link{stations}}. Default is NULL. If NULL, a data frame
of 30-year normals for all stations will be returned.}

\item{tidy}{Logical. Returns tidy data. Default is TRUE.
\describe{
\item{TRUE}{Returns a tidy data frame of temperature (tmax and tmin)
and precipitation data. This option drops seasonal earliest/last
dates for temperatures and number of days above/below
temperatures.}
\item{FALSE}{Returns the raw dat from ClimateAnalyzer.org.}
}}
}
\value{
A \code{\link[tibble]{tibble}}.
}
\description{
This function imports 30-year normals calculated by NOAA from
\href{http://www.climateanalyzer.org/}{ClimateAnalyzer.org} into R.
}
\examples{
library(climateAnalyzeR)

# Import 1981-2010 normals. Default.
normals(station_id = "capitol_reef_np")

# Import 1971-2000 normals for multiple stations
normals(ref_period = "1971-2000",
        station_id = c("dinosaur_nm", "dinosaur_quarry_area"))

}
