% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderSummaries.R
\name{renderSummary}
\alias{renderSummary}
\title{Render reports}
\usage{
renderSummary(
  my_report,
  station_id,
  station_name = NULL,
  my_year = NULL,
  my_dir = NULL
)
}
\arguments{
\item{my_report}{Type of report to render. The options are a calendar
year ("calendar_year") and a water year ("water_year") report.}

\item{station_id}{The character string of the \emph{station_id} field from
\code{\link{stations}}..}

\item{station_name}{Optional. A string to be use for the station name in the
report. Default is NULL. If NULL the station name will be pulled from
\code{\link{stations}}. This is displayed in several location in the
report and the default name can be changed.}

\item{my_year}{Optional. The four-digit year of interest. Default is NULL. The
current year will be used if left as NULL.}

\item{my_dir}{Optional. String of directory path to save report to. Default
is NULL. If NULL, the report is saved to the your desktop.}
}
\value{
A PDF.
}
\description{
This function will render a PDF report from an RMarkdown (Rmd) script.
}
\examples{
\dontrun{
#' library(climateAnalyzeR)

# Render water year report for current water year
renderSummary(my_report = "water_year", station_id = "hans_flat_rs")

# Render calendar year report for 2018 with custom station name
renderSummary(my_report = "calendar_year", station_id = "canyonlands_theneck",
              station_name = "Island in the Sky, Canyonlands National Park",
              my_year = 2018)
}
}
