% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderSummaries.R
\name{renderSummary}
\alias{renderSummary}
\title{Render reports}
\usage{
renderSummary(
  park_name,
  water_yr,
  my_park,
  report = "Water Year",
  my_dir = NULL
)
}
\arguments{
\item{park_name}{The full name of the park unit of interest.}

\item{water_yr}{The four-digit water year you want to summarize.}

\item{my_park}{The four-letter park code for the park of interest.}

\item{report}{Type of report to render. Currently the only option is "Water Year".}

\item{my_dir}{Direectory to save repor to. If NULL, the report is saved to the your desktop.}
}
\value{

}
\description{
This funciton will render a PDF report from an RMarkdown (Rmd) script.
}
\examples{
# Render a report for Arches National Park for the 2020 water year.
renderSummary("Arches National Park", 2020, "ARCH", report = "Water Year")
}
