% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{stations}
\alias{stations}
\title{Import station informaiton from ClimateAnalyzer.org}
\usage{
stations(my_stations = NULL)
}
\arguments{
\item{my_stations}{A string of the station name. This is used as an argument
for dplyr::filter to reduce the data frame to the desired stations.
Default is NULL. If NULL, a data frame of all stations is returned.}
}
\value{
A \code{\link[tibble:tibble]{tibble}}.
}
\description{
This function imports weather station information from
\href{http://www.climateanalyzer.org/}{ClimateAnalyzer.org} into R.
}
\examples{
# Import all information for all stations.
stations()

# Filter station by name
stations(my_stations = "arches")
stations(my_stations = c("Canyonlands The Neck", "Canyonlands The Needle", "Hans Flat RS"))
}
