% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{stations}
\alias{stations}
\title{Import station informaiton from ClimateAnalyzer.org}
\usage{
stations(station_name = NULL)
}
\arguments{
\item{station_name}{A string of the station name. This is used as an argument
for dplyr::filter to reduce the data frame to the desired stations.
Default is NULL. If NULL, a data frame of all stations is returned.}
}
\value{
A \code{\link[tibble:tibble]{tibble}}.
}
\description{
Import station informaiton from ClimateAnalyzer.org
}
\examples{
# Import all information for all stations.
stations()

# Filter station by name
stations(station_name = "Arches NP")
stations(station_name = c("Canyonlands The Neck", "Canyonlands The Needle", "Hans Flat RS"))
}
