% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{stations}
\alias{stations}
\title{Import station information}
\usage{
stations(my_name = NULL, my_stations = NULL, my_id = NULL)
}
\arguments{
\item{my_name}{A string of the station name. This is used to reduce the data
frame using \code{\link[dplyr::filter]{filter}}. Default is NULL.}

\item{my_stations}{A string of the station id. This is used to reduce the data
frame using \code{\link[dplyr::filter]{filter}}. Default is NULL.}

\item{my_id}{A string of the station Id. This is used to reduce the data
frame using \code{\link[dplyr::filter]{filter}}. Default is NULL.}
}
\value{
A \code{\link[tibble:tibble]{tibble}}.
}
\description{
This function imports weather station information from
\href{http://www.climateanalyzer.org/}{ClimateAnalyzer.org} into R.
}
\examples{
library(climateAnalyzeR)

# Import all information for all stations.
stations()

# Filter station by name
stations(my_stations = "arches")
stations(my_id = c("426053", "424100"))

}
