% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{stations}
\alias{stations}
\title{Import station information}
\usage{
stations(station_name = NULL, station_id = NULL, id = NULL)
}
\arguments{
\item{station_name}{A string of the station \emph{name}. This is used to reduce
the data frame using \code{\link[dplyr]{filter}}. Default is NULL.}

\item{station_id}{A string of the \emph{station_id}. This is used to reduce
the data frame using \code{\link[dplyr]{filter}}. Default is NULL.}

\item{id}{A string of the station \emph{id}. This is used to reduce the
data frame using \code{\link[dplyr]{filter}}. Default is NULL.}
}
\value{
A \code{\link[tibble:tibble]{tibble}}.
}
\description{
This function imports weather station information from
\href{http://www.climateanalyzer.org/}{ClimateAnalyzer.org} into R.
}
\examples{
library(climateAnalyzeR)

# Import all information for all stations.
stations()

# Import station inforation using the name of the station
stations(station_name = "Black Canyon of the Gunnison")

# Import information for multiple stations using station_id
stations(station_id = c("blue_mesa_lake", "black_canyon_of_the_gunnison"))

# Import information for multiple stations using id numbers
stations(id = c(50754, 50797))

}
